sdk: BREAKING CHANGE: File class moved to models.files
