+++ sdk/tests/e2e_examples/check_build_acceptance.py
@@ -23,9 +23,9 @@ from typing import TYPE_CHECKING
 
 from spectrumx import Client
 from spectrumx import enable_logging
-from spectrumx.models import Capture
-from spectrumx.models import Dataset
-from spectrumx.models import File
+from spectrumx.models.captures import Capture
+from spectrumx.models.files import File
+from spectrumx.models.datasets import Dataset
 
 NOT_IMPLEMENTED = "This example is not yet implemented."
